Refactoring. Mainly reworked Command classes.
Reworked almost all Command classes.
Added comments, cleaned code, reworked tough places.

+++ src/main/java/ua.training.tts/controller/command/admin/project/ProjectDelete.java
@@ -1,4 +1,4 @@
-package ua.training.tts.controller.command.admin;
+package ua.training.tts.controller.command.admin.project;
 
 import ua.training.tts.constant.ReqSesParameters;
 import ua.training.tts.constant.controller.Servlet;
@@ -9,7 +9,11 @@ import ua.training.tts.model.entity.Employee;
 import ua.training.tts.model.service.ProjectService;
 
 import javax.servlet.http.HttpServletRequest;
+import java.util.Objects;
 
+/**
+ * Tries to delete selected project.
+ */
 @AccessRights(acceptedRoles = {Employee.AccountRole.ADMIN}, isAvailableForGuests = false)
 public class ProjectDelete implements Command {
 
@@ -21,8 +25,10 @@ public class ProjectDelete implements Command {
 
     @Override
     public String execute(HttpServletRequest request) {
-        Integer id = Integer.parseInt(request.getParameter(ReqSesParameters.PROJECT_ID));
-        service.deleteById(id);
-        return Servlet.REDIRECT + CommandParameters.MAIN;
+        if (Objects.nonNull(request.getParameter(ReqSesParameters.PROJECT_ID))) {
+            Integer id = Integer.parseInt(request.getParameter(ReqSesParameters.PROJECT_ID));
+            service.deleteById(id);
+        }
+        return CommandParameters.REDIRECT + CommandParameters.ADMIN_ALL_PROJECTS;
     }
 }
