+++ src/main/java/ua.training.tts/controller/command/admin/task/TaskDelete.java
@@ -1,8 +1,6 @@
-package ua.training.tts.controller.command.admin;
+package ua.training.tts.controller.command.admin.task;
 
 import ua.training.tts.constant.ReqSesParameters;
-import ua.training.tts.constant.controller.FilterParameters;
-import ua.training.tts.constant.controller.Servlet;
 import ua.training.tts.constant.controller.command.CommandParameters;
 import ua.training.tts.controller.command.Command;
 import ua.training.tts.controller.util.AccessRights;
@@ -10,7 +8,11 @@ import ua.training.tts.model.entity.Employee;
 import ua.training.tts.model.service.TaskService;
 
 import javax.servlet.http.HttpServletRequest;
+import java.util.Objects;
 
+/**
+ * Tries to delete selected task.
+ */
 @AccessRights(acceptedRoles = {Employee.AccountRole.ADMIN}, isAvailableForGuests = false)
 public class TaskDelete implements Command {
 
@@ -22,8 +24,10 @@ public class TaskDelete implements Command {
 
     @Override
     public String execute(HttpServletRequest request) {
-        Integer id = Integer.parseInt(request.getParameter(ReqSesParameters.TASK_ID));
-        service.deleteById(id);
-        return Servlet.REDIRECT + CommandParameters.MAIN;
+        if (Objects.nonNull(request.getParameter(ReqSesParameters.TASK_ID))) {
+            Integer id = Integer.parseInt(request.getParameter(ReqSesParameters.TASK_ID));
+            service.deleteById(id);
+        }
+        return CommandParameters.REDIRECT + CommandParameters.MAIN;
     }
 }
